refactor: extract pending settings trait
