+++ nuclear-engagement/includes/SettingsRepository.php
@@ -122,6 +122,7 @@ final class SettingsRepository
     }
 
     use SettingsAccessTrait;
+    use PendingSettingsTrait;
 
 
     /* ===================================================================
@@ -207,40 +208,6 @@ final class SettingsRepository
         return array_key_exists($key, $all);
     }
 
-    /**
-     * Remove a setting.
-     */
-    public function remove(string $key): self {
-        if (isset($this->pending[$key])) {
-            unset($this->pending[$key]);
-        }
-        
-        // Set to null to indicate removal
-        $this->pending[$key] = null;
-        return $this;
-    }
-
-    /**
-     * Check if there are pending changes.
-     */
-    public function has_pending(): bool {
-        return !empty($this->pending);
-    }
-
-    /**
-     * Get all pending changes.
-     */
-    public function get_pending(): array {
-        return $this->pending;
-    }
-
-    /**
-     * Clear pending changes without saving.
-     */
-    public function clear_pending(): self {
-        $this->pending = [];
-        return $this;
-    }
 
     /**
      * Determine if settings should be autoloaded.
